Samsung Open Source: G9550ZCU3CRK2 OO

+++ crypto/first_file.c
@@ -28,7 +28,7 @@ void first_crypto_text (void)
 
 #define KERNEL_KASLR_16K_ALGIN
 #define DDR_START_FIRST		(0x80000000)
-#define DDR_START_SECOND	(0x40000000)
+#define DDR_START_SECOND	(0x40000000) // Used only for MSM8998 6GB RAM
 
 #ifdef  KERNEL_KASLR_16K_ALGIN
 #define KASLR_FIRST_SLOT	(0x80000)
@@ -45,13 +45,13 @@ get_builtime_crypto_hmac (void)
 	u64 *kernel_addr = (u64 *) &__boot_kernel_offset;
 	u64 offset = 0;
 	u64 idx = 0;
-	if (ddr_start_type) {
+	if (ddr_start_type) { // This value is set only for MSM8998
 		if (ddr_start_type == 1)
-			offset = (u64)kernel_addr[1] + (u64)kernel_addr[0] - (u64)KASLR_FIRST_SLOT - (u64)DDR_START_FIRST;
+			offset = (u64)((u64)kernel_addr[1] + (u64)kernel_addr[0] - (u64)KASLR_FIRST_SLOT - (u64)DDR_START_FIRST);
 		else
-			offset = (u64)kernel_addr[1] + (u64)kernel_addr[0] - (u64)KASLR_FIRST_SLOT - (u64)DDR_START_SECOND;
+			offset = (u64)((u64)kernel_addr[1] + (u64)kernel_addr[0] - (u64)KASLR_FIRST_SLOT - (u64)DDR_START_SECOND);
 	} else {
-		offset = (u64)kernel_addr[1] + (u64)kernel_addr[0] - (u64)KASLR_FIRST_SLOT - (u64)DDR_START_FIRST;
+		offset = (u64)((u64)kernel_addr[1] + (u64)kernel_addr[0] - (u64)KASLR_FIRST_SLOT - (u64)DDR_START_FIRST);
 	}
 
 	idx = (offset / KASLR_ALIGN);
